hospital_records/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
boto3==1.17.87
botocore==1.20.87
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
Django==3.2.6
django-betterforms==1.2
django-six==1.0.4
django-toggle-switch-widget==0.1.1
django-utils==0.0.2
django-utils-six==2.0
idna==2.10
jmespath==0.10.0
mysql-connector-python==8.0.26
mysqlclient==2.0.3
psycopg2==2.9.1
pymssql==2.2.2
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
